feat: implement provider-specific playback and auth handling for Jellyfin and Plex
